requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
dirtyjson==1.0.8
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-heroku==0.3.1
djangorestframework==3.14.0
frozenlist==1.4.0
gevent==23.9.1
greenlet==3.0.0rc3
gunicorn==21.2.0
idna==3.4
ijson==3.2.3
multidict==6.0.4
newrelic==9.0.0
openai==0.27.8
packaging==23.1
psycogreen==1.0.2
psycopg==3.1.10
psycopg-binary==3.1.10
pytz==2023.3
requests==2.31.0
sentry-sdk==1.30.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
whitenoise==6.5.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
